Magento 2 - refactoring

+++ Controller/Adminhtml/Ecommerce/Delete.php
@@ -7,8 +7,7 @@ namespace GetResponse\GetResponseIntegration\Controller\Adminhtml\Ecommerce;
 use Exception;
 use GetResponse\GetResponseIntegration\Controller\Adminhtml\AbstractController;
 use GetResponse\GetResponseIntegration\Domain\GetResponse\Api\ApiClientFactory;
-use GetResponse\GetResponseIntegration\Domain\SharedKernel\Scope;
-use GetResponse\GetResponseIntegration\Helper\MagentoStore;
+use GetResponse\GetResponseIntegration\Domain\SharedKernel\Exception\IdNotFoundException;
 use GetResponse\GetResponseIntegration\Helper\Message;
 use GetResponse\GetResponseIntegration\Helper\Route;
 use GrShareCode\Shop\Command\DeleteShopCommand;
@@ -19,39 +18,34 @@ class Delete extends AbstractController
 {
     private $apiClient;
 
-    public function __construct(
-        Context $context,
-        ApiClientFactory $apiClientFactory,
-        MagentoStore $magentoStore
-    ) {
+    public function __construct(Context $context, ApiClientFactory $apiClientFactory)
+    {
         parent::__construct($context);
 
-        $this->apiClient = $apiClientFactory->createGetResponseApiClient(
-            new Scope($magentoStore->getStoreIdFromUrl())
-        );
+        $this->apiClient = $apiClientFactory->createGetResponseApiClient($this->scope);
     }
 
     public function execute()
     {
+        parent::execute();
+
+        if (!$this->isConnected()) {
+            return $this->redirectToStore(Route::ACCOUNT_INDEX_ROUTE);
+        }
+
         try {
-            $id = $this->getRequest()->getParam('id');
+            $id = $this->request->getParam('id');
 
             if (empty($id)) {
-                throw new Exception(Message::INCORRECT_SHOP);
+                throw new IdNotFoundException(Message::INCORRECT_SHOP);
             }
 
             $service = new ShopService($this->apiClient);
             $service->deleteShop(new DeleteShopCommand($id));
 
-            $resultRedirect = $this->resultRedirectFactory->create();
-            $resultRedirect->setPath(Route::ECOMMERCE_INDEX_ROUTE);
-
-            return $resultRedirect;
+            return $this->redirect($this->_redirect->getRefererUrl(), Message::SHOP_DELETED);
         } catch (Exception $e) {
-            $resultRedirect = $this->resultRedirectFactory->create();
-            $this->messageManager->addErrorMessage($e->getMessage());
-            $resultRedirect->setPath(Route::ECOMMERCE_INDEX_ROUTE);
-            return $resultRedirect;
+            return $this->redirect($this->_redirect->getRefererUrl(),$e->getMessage(), true);
         }
     }
 }
